Improve code quality #86
